requirements.txt
torch
torchvision
matplotlib
pyyaml